+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Hacker News IDEA
 
 ## [Unreleased]
+
+### Added
+
+- `{by}` and `{id}` format specifiers.
+
+### Fixed
+
+- Change "Top Stories on Hacker News" to "Top Stories" as menu item title for popup.
+- Use java.net.http to fetch data, which allows specifying user agent.
+
 ## [1.3.0]
 
 ### Added